# backend-v2/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Generated Backend API",
    "description": "API documentation generated by Backend Generator v2",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AchievementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "difficulty": {
            "type": "string"
          },
          "requirements": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "category",
          "icon",
          "points",
          "difficulty",
          "requirements",
          "created_at",
          "updated_at"
        ]
      },
      "CreateAchievementRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "difficulty": {
            "type": "string"
          },
          "requirements": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "category",
          "icon",
          "points",
          "difficulty",
          "requirements"
        ]
      },
...
      "ContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "parent_id": {
            "type": "string"
          },
          "order_index": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "content_body": {
            "type": "object"
          },
          "difficulty_level": {
            "type": "string"
          },
          "estimated_duration": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "objectives": {
            "type": "object"
          },
          "prerequisites": {
            "type": "object"
          },
          "is_published": {
            "type": "boolean"
          },
          "version": {
            "type": "number"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content_type",
          "parent_id",
          "order_index",
          "description",
          "content_body",
          "difficulty_level",
          "estimated_duration",
          "category",
          "objectives",
          "prerequisites",
          "is_published",
          "version",
          "created_at",
          "updated_at"
        ]
      },
      "CreateContentRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content_type": {
            "type": "string"
          },
          "parent_id": {
            "type": "string"
          },
          "order_index": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "content_body": {
            "type": "object"
          },
          "difficulty_level": {
            "type": "string"
          },
          "estimated_duration": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "objectives": {
            "type": "object"
          },
          "prerequisites": {
            "type": "object"
          },
          "is_published": {
            "type": "boolean"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "title",
          "content_type",
          "parent_id",
          "order_index",
          "content_body",
          "difficulty_level",
          "estimated_duration",
          "category",
          "objectives",
          "prerequisites",
          "is_published",
          "version"
        ]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
      "ValidationError": {
        "description": "Request validation failed"
      },
      "InternalServerError": {
        "description": "Internal server error"
      }
    }
  },
  "paths": {
    "/api/achievements/": {
      "post": {
        "summary": "Create new achievement",
        "tags": [
          "Achievement"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AchievementResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAchievementRequest"
              }
            }
          }
        }
      },
      "get": {
        "summary": "listAchievements - Achievement operation",
        "tags": [
          "Achievement"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AchievementResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/achievements/:id": {
      "get": {
        "summary": "Get achievement information",
        "tags": [
          "Achievement"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AchievementResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update achievement",
        "tags": [
          "Achievement"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AchievementResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAchievementRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete achievement",
        "tags": [
          "Achievement"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AchievementResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/content/": {
      "post": {
        "summary": "Create new content",
        "tags": [
          "Content"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentRequest"
              }
            }
          }
        }
      },
      "get": {
        "summary": "listContents - Content operation",
        "tags": [
          "Content"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/content/:id": {
      "get": {
        "summary": "Get content information",
        "tags": [
          "Content"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update content",
        "tags": [
          "Content"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete content",
        "tags": [
          "Content"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}